Crate mlx_sys

Re-exports

pub use self::mlx_array_dtype_ as mlx_array_dtype;

pub use self::mlx_device_type_ as mlx_device_type;

Structs

_OSUnalignedU16

_OSUnalignedU32

_OSUnalignedU64

__BindgenBitfieldUnit

__BindgenComplex

__BindgenFloat16

__arm_legacy_debug_state

__arm_pagein_state

__darwin_arm_cpmu_state64

__darwin_arm_debug_state32

__darwin_arm_debug_state64

__darwin_arm_exception_state

__darwin_arm_exception_state64

__darwin_arm_neon_state

__darwin_arm_neon_state64

__darwin_arm_thread_state

__darwin_arm_thread_state64

__darwin_arm_vfp_state

__darwin_mcontext32

__darwin_mcontext64

__darwin_pthread_handler_rec

__darwin_sigaltstack

__darwin_ucontext

__sFILE

__sFILEX

__sbuf

__sigaction

__siginfo

_malloc_zone_t

_opaque_pthread_attr_t

_opaque_pthread_cond_t

_opaque_pthread_condattr_t

_opaque_pthread_mutex_t

_opaque_pthread_mutexattr_t

_opaque_pthread_once_t

_opaque_pthread_rwlock_t

_opaque_pthread_rwlockattr_t

_opaque_pthread_t

div_t

ldiv_t

lldiv_t

mlx_array_

mlx_closure_

mlx_closure_value_and_grad_

mlx_device_

mlx_distributed_group_

mlx_future_

mlx_map_string_to_array_

mlx_map_string_to_array_iterator_

mlx_map_string_to_string_

mlx_map_string_to_string_iterator_

mlx_safetensors_

mlx_stream_

mlx_string_

mlx_vector_array_

mlx_vector_vector_array_

proc_rlimit_control_wakeupmon

rlimit

rusage

rusage_info_v0

rusage_info_v1

rusage_info_v2

rusage_info_v3

rusage_info_v4

rusage_info_v5

rusage_info_v6

sigaction

sigevent

sigstack

sigvec

timeval

wait__bindgen_ty_1

wait__bindgen_ty_2

Constants

BIG_ENDIAN

BUFSIZ

BUS_ADRALN

BUS_ADRERR

BUS_NOOP

BUS_OBJERR

BYTE_ORDER

CLD_CONTINUED

CLD_DUMPED

CLD_EXITED

CLD_KILLED

CLD_NOOP

CLD_STOPPED

CLD_TRAPPED

CPUMON_MAKE_FATAL

DBL_HAS_SUBNORM

EOF

EXIT_FAILURE

EXIT_SUCCESS

FILENAME_MAX

FLT_HAS_SUBNORM

FOOTPRINT_INTERVAL_RESET

FOPEN_MAX

FPE_FLTDIV

FPE_FLTINV

FPE_FLTOVF

FPE_FLTRES

FPE_FLTSUB

FPE_FLTUND

FPE_INTDIV

FPE_INTOVF

FPE_NOOP

ILL_BADSTK

ILL_COPROC

ILL_ILLADR

ILL_ILLOPC

ILL_ILLOPN

ILL_ILLTRP

ILL_NOOP

ILL_PRVOPC

ILL_PRVREG

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INT64_MAX

INT64_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_FAST64_MAX

INT_FAST64_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

INT_LEAST64_MAX

INT_LEAST64_MIN

IOPOL_APPLICATION

IOPOL_ATIME_UPDATES_DEFAULT

IOPOL_ATIME_UPDATES_OFF

IOPOL_DEFAULT

IOPOL_IMPORTANT

IOPOL_MATERIALIZE_DATALESS_FILES_DEFAULT

IOPOL_MATERIALIZE_DATALESS_FILES_OFF

IOPOL_MATERIALIZE_DATALESS_FILES_ON

IOPOL_NORMAL

IOPOL_PASSIVE

IOPOL_SCOPE_DARWIN_BG

IOPOL_SCOPE_PROCESS

IOPOL_SCOPE_THREAD

IOPOL_STANDARD

IOPOL_THROTTLE

IOPOL_TYPE_DISK

IOPOL_TYPE_VFS_ALLOW_LOW_SPACE_WRITES

IOPOL_TYPE_VFS_ATIME_UPDATES

IOPOL_TYPE_VFS_DISALLOW_RW_FOR_O_EVTONLY

IOPOL_TYPE_VFS_IGNORE_CONTENT_PROTECTION

IOPOL_TYPE_VFS_IGNORE_PERMISSIONS

IOPOL_TYPE_VFS_MATERIALIZE_DATALESS_FILES

IOPOL_TYPE_VFS_SKIP_MTIME_UPDATE

IOPOL_TYPE_VFS_STATFS_NO_DATA_VOLUME

IOPOL_TYPE_VFS_TRIGGER_RESOLVE

IOPOL_UTILITY

IOPOL_VFS_ALLOW_LOW_SPACE_WRITES_OFF

IOPOL_VFS_ALLOW_LOW_SPACE_WRITES_ON

IOPOL_VFS_CONTENT_PROTECTION_DEFAULT

IOPOL_VFS_CONTENT_PROTECTION_IGNORE

IOPOL_VFS_DISALLOW_RW_FOR_O_EVTONLY_DEFAULT

IOPOL_VFS_DISALLOW_RW_FOR_O_EVTONLY_ON

IOPOL_VFS_IGNORE_PERMISSIONS_OFF

IOPOL_VFS_IGNORE_PERMISSIONS_ON

IOPOL_VFS_NOCACHE_WRITE_FS_BLKSIZE_DEFAULT

IOPOL_VFS_NOCACHE_WRITE_FS_BLKSIZE_ON

IOPOL_VFS_SKIP_MTIME_UPDATE_OFF

IOPOL_VFS_SKIP_MTIME_UPDATE_ON

IOPOL_VFS_STATFS_FORCE_NO_DATA_VOLUME

IOPOL_VFS_STATFS_NO_DATA_VOLUME_DEFAULT

IOPOL_VFS_TRIGGER_RESOLVE_DEFAULT

IOPOL_VFS_TRIGGER_RESOLVE_OFF

LDBL_HAS_SUBNORM

LITTLE_ENDIAN

L_ctermid

L_tmpnam

MAC_OS_VERSION_11_0

MAC_OS_VERSION_11_1

MAC_OS_VERSION_11_3

MAC_OS_VERSION_11_4

MAC_OS_VERSION_11_5

MAC_OS_VERSION_11_6

MAC_OS_VERSION_12_0

MAC_OS_VERSION_12_1

MAC_OS_VERSION_12_2

MAC_OS_VERSION_12_3

MAC_OS_VERSION_12_4

MAC_OS_VERSION_12_5

MAC_OS_VERSION_12_6

MAC_OS_VERSION_12_7

MAC_OS_VERSION_13_0

MAC_OS_VERSION_13_1

MAC_OS_VERSION_13_2

MAC_OS_VERSION_13_3

MAC_OS_VERSION_13_4

MAC_OS_VERSION_13_5

MAC_OS_VERSION_13_6

MAC_OS_VERSION_14_0

MAC_OS_VERSION_14_1

MAC_OS_VERSION_14_2

MAC_OS_VERSION_14_3

MAC_OS_VERSION_14_4

MAC_OS_VERSION_14_5

MAC_OS_X_VERSION_10_0

MAC_OS_X_VERSION_10_1

MAC_OS_X_VERSION_10_2

MAC_OS_X_VERSION_10_3

MAC_OS_X_VERSION_10_4

MAC_OS_X_VERSION_10_5

MAC_OS_X_VERSION_10_6

MAC_OS_X_VERSION_10_7

MAC_OS_X_VERSION_10_8

MAC_OS_X_VERSION_10_9

MAC_OS_X_VERSION_10_10

MAC_OS_X_VERSION_10_10_2

MAC_OS_X_VERSION_10_10_3

MAC_OS_X_VERSION_10_11

MAC_OS_X_VERSION_10_11_2

MAC_OS_X_VERSION_10_11_3

MAC_OS_X_VERSION_10_11_4

MAC_OS_X_VERSION_10_12

MAC_OS_X_VERSION_10_12_1

MAC_OS_X_VERSION_10_12_2

MAC_OS_X_VERSION_10_12_4

MAC_OS_X_VERSION_10_13

MAC_OS_X_VERSION_10_13_1

MAC_OS_X_VERSION_10_13_2

MAC_OS_X_VERSION_10_13_4

MAC_OS_X_VERSION_10_14

MAC_OS_X_VERSION_10_14_1

MAC_OS_X_VERSION_10_14_4

MAC_OS_X_VERSION_10_14_5

MAC_OS_X_VERSION_10_14_6

MAC_OS_X_VERSION_10_15

MAC_OS_X_VERSION_10_15_1

MAC_OS_X_VERSION_10_15_4

MAC_OS_X_VERSION_10_16

MINSIGSTKSZ

NSIG

PDP_ENDIAN

POLL_ERR

POLL_HUP

POLL_IN

POLL_MSG

POLL_OUT

POLL_PRI

PRIO_DARWIN_BG

PRIO_DARWIN_NONUI

PRIO_DARWIN_PROCESS

PRIO_DARWIN_THREAD

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

P_tmpdir

RAND_MAX

RENAME_EXCL

RENAME_NOFOLLOW_ANY

RENAME_RESERVED1

RENAME_SECLUDE

RENAME_SWAP

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_CPU_USAGE_MONITOR

RLIMIT_DATA

RLIMIT_FOOTPRINT_INTERVAL

RLIMIT_FSIZE

RLIMIT_MEMLOCK

RLIMIT_NOFILE

RLIMIT_NPROC

RLIMIT_RSS

RLIMIT_STACK

RLIMIT_THREAD_CPULIMITS

RLIMIT_WAKEUPS_MONITOR

RLIM_NLIMITS

RSIZE_MAX

RUSAGE_CHILDREN

RUSAGE_INFO_CURRENT

RUSAGE_INFO_V0

RUSAGE_INFO_V1

RUSAGE_INFO_V2

RUSAGE_INFO_V3

RUSAGE_INFO_V4

RUSAGE_INFO_V5

RUSAGE_INFO_V6

RUSAGE_SELF

RU_PROC_RUNS_RESLIDE

SA_64REGSET

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SA_USERSPACE_MASK

SA_USERTRAMP

SEEK_CUR

SEEK_DATA

SEEK_END

SEEK_HOLE

SEEK_SET

SEGV_ACCERR

SEGV_MAPERR

SEGV_NOOP

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCONT

SIGEMT

SIGEV_NONE

SIGEV_SIGNAL

SIGEV_THREAD

SIGFPE

SIGHUP

SIGILL

SIGINFO

SIGINT

SIGIO

SIGIOT

SIGKILL

SIGPIPE

SIGPROF

SIGQUIT

SIGSEGV

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIG_BLOCK

SIG_SETMASK

SIG_UNBLOCK

SIZE_MAX

SI_ASYNCIO

SI_MESGQ

SI_QUEUE

SI_TIMER

SI_USER

SS_DISABLE

SS_ONSTACK

SV_INTERRUPT

SV_NOCLDSTOP

SV_NODEFER

SV_ONSTACK

SV_RESETHAND

SV_SIGINFO

TMP_MAX

TRAP_BRKPT

TRAP_TRACE

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINT64_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_FAST64_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

UINT_LEAST64_MAX

WAIT_ANY

WAIT_MYPGRP

WAKEMON_DISABLE

WAKEMON_ENABLE

WAKEMON_GET_PARAMS

WAKEMON_MAKE_FATAL

WAKEMON_SET_DEFAULTS

WCONTINUED

WCOREFLAG

WEXITED

WINT_MAX

WINT_MIN

WNOHANG

WNOWAIT

WSTOPPED

WUNTRACED

_ARM_SIGNAL_

_DARWIN_FEATURE_64_BIT_INODE

_DARWIN_FEATURE_ONLY_64_BIT_INODE

_DARWIN_FEATURE_ONLY_UNIX_CONFORMANCE

_DARWIN_FEATURE_ONLY_VERS_1050

_DARWIN_FEATURE_UNIX_CONFORMANCE

_FORTIFY_SOURCE

_IOFBF

_IOLBF

_IONBF

_QUAD_HIGHWORD

_QUAD_LOWWORD

_RLIMIT_POSIX_FLAG

_USE_FORTIFY_LEVEL

_WSTOPPED

__API_TO_BE_DEPRECATED

__API_TO_BE_DEPRECATED_DRIVERKIT

__API_TO_BE_DEPRECATED_IOS

__API_TO_BE_DEPRECATED_MACCATALYST

__API_TO_BE_DEPRECATED_MACOS

__API_TO_BE_DEPRECATED_TVOS

__API_TO_BE_DEPRECATED_VISIONOS

__API_TO_BE_DEPRECATED_WATCHOS

__BRIDGEOS_2_0

__BRIDGEOS_3_0

__BRIDGEOS_3_1

__BRIDGEOS_3_4

__BRIDGEOS_4_0

__BRIDGEOS_4_1

__BRIDGEOS_5_0

__BRIDGEOS_5_1

__BRIDGEOS_5_3

__BRIDGEOS_6_0

__BRIDGEOS_6_2

__BRIDGEOS_6_4

__BRIDGEOS_6_5

__BRIDGEOS_6_6

__BRIDGEOS_7_0

__BRIDGEOS_7_1

__BRIDGEOS_7_2

__BRIDGEOS_7_3

__BRIDGEOS_7_4

__BRIDGEOS_7_6

__BRIDGEOS_8_0

__BRIDGEOS_8_1

__BRIDGEOS_8_2

__BRIDGEOS_8_3

__BRIDGEOS_8_4

__BRIDGEOS_8_5

__DARWIN_64_BIT_INO_T

__DARWIN_BIG_ENDIAN

__DARWIN_BYTE_ORDER

__DARWIN_C_ANSI

__DARWIN_C_FULL

__DARWIN_C_LEVEL

__DARWIN_LITTLE_ENDIAN

__DARWIN_NON_CANCELABLE

__DARWIN_NO_LONG_LONG

__DARWIN_NSIG

__DARWIN_ONLY_64_BIT_INO_T

__DARWIN_ONLY_UNIX_CONFORMANCE

__DARWIN_ONLY_VERS_1050

__DARWIN_OPAQUE_ARM_THREAD_STATE64

__DARWIN_PDP_ENDIAN

__DARWIN_SUF_EXTSN

__DARWIN_UNIX03

__DARWIN_VERS_1050

__DARWIN_WCHAR_MIN

__DRIVERKIT_19_0

__DRIVERKIT_20_0

__DRIVERKIT_21_0

__DRIVERKIT_22_0

__DRIVERKIT_22_4

__DRIVERKIT_22_5

__DRIVERKIT_22_6

__DRIVERKIT_23_0

__DRIVERKIT_23_1

__DRIVERKIT_23_2

__DRIVERKIT_23_3

__DRIVERKIT_23_4

__DRIVERKIT_23_5

__ENABLE_LEGACY_MAC_AVAILABILITY

__IPHONE_2_0

__IPHONE_2_1

__IPHONE_2_2

__IPHONE_3_0

__IPHONE_3_1

__IPHONE_3_2

__IPHONE_4_0

__IPHONE_4_1

__IPHONE_4_2

__IPHONE_4_3

__IPHONE_5_0

__IPHONE_5_1

__IPHONE_6_0

__IPHONE_6_1

__IPHONE_7_0

__IPHONE_7_1

__IPHONE_8_0

__IPHONE_8_1

__IPHONE_8_2

__IPHONE_8_3

__IPHONE_8_4

__IPHONE_9_0

__IPHONE_9_1

__IPHONE_9_2

__IPHONE_9_3

__IPHONE_10_0

__IPHONE_10_1

__IPHONE_10_2

__IPHONE_10_3

__IPHONE_11_0

__IPHONE_11_1

__IPHONE_11_2

__IPHONE_11_3

__IPHONE_11_4

__IPHONE_12_0

__IPHONE_12_1

__IPHONE_12_2

__IPHONE_12_3

__IPHONE_12_4

__IPHONE_13_0

__IPHONE_13_1

__IPHONE_13_2

__IPHONE_13_3

__IPHONE_13_4

__IPHONE_13_5

__IPHONE_13_6

__IPHONE_13_7

__IPHONE_14_0

__IPHONE_14_1

__IPHONE_14_2

__IPHONE_14_3

__IPHONE_14_4

__IPHONE_14_5

__IPHONE_14_6

__IPHONE_14_7

__IPHONE_14_8

__IPHONE_15_0

__IPHONE_15_1

__IPHONE_15_2

__IPHONE_15_3

__IPHONE_15_4

__IPHONE_15_5

__IPHONE_15_6

__IPHONE_15_7

__IPHONE_15_8

__IPHONE_16_0

__IPHONE_16_1

__IPHONE_16_2

__IPHONE_16_3

__IPHONE_16_4

__IPHONE_16_5

__IPHONE_16_6

__IPHONE_16_7

__IPHONE_17_0

__IPHONE_17_1

__IPHONE_17_2

__IPHONE_17_3

__IPHONE_17_4

__IPHONE_17_5

__MAC_10_0

__MAC_10_1

__MAC_10_2

__MAC_10_3

__MAC_10_4

__MAC_10_5

__MAC_10_6

__MAC_10_7

__MAC_10_8

__MAC_10_9

__MAC_10_10

__MAC_10_10_2

__MAC_10_10_3

__MAC_10_11

__MAC_10_11_2

__MAC_10_11_3

__MAC_10_11_4

__MAC_10_12

__MAC_10_12_1

__MAC_10_12_2

__MAC_10_12_4

__MAC_10_13

__MAC_10_13_1

__MAC_10_13_2

__MAC_10_13_4

__MAC_10_14

__MAC_10_14_1

__MAC_10_14_4

__MAC_10_14_5

__MAC_10_14_6

__MAC_10_15

__MAC_10_15_1

__MAC_10_15_4

__MAC_10_16

__MAC_11_0

__MAC_11_1

__MAC_11_3

__MAC_11_4

__MAC_11_5

__MAC_11_6

__MAC_12_0

__MAC_12_1

__MAC_12_2

__MAC_12_3

__MAC_12_4

__MAC_12_5

__MAC_12_6

__MAC_12_7

__MAC_13_0

__MAC_13_1

__MAC_13_2

__MAC_13_3

__MAC_13_4

__MAC_13_5

__MAC_13_6

__MAC_14_0

__MAC_14_1

__MAC_14_2

__MAC_14_3

__MAC_14_4

__MAC_14_5

__MAC_OS_X_VERSION_MAX_ALLOWED

__PTHREAD_ATTR_SIZE__

__PTHREAD_CONDATTR_SIZE__

__PTHREAD_COND_SIZE__

__PTHREAD_MUTEXATTR_SIZE__

__PTHREAD_MUTEX_SIZE__

__PTHREAD_ONCE_SIZE__

__PTHREAD_RWLOCKATTR_SIZE__

__PTHREAD_RWLOCK_SIZE__

__PTHREAD_SIZE__

__SALC

__SAPP

__SEOF

__SERR

__SIGN

__SLBF

__SMBF

__SMOD

__SNBF

__SNPT

__SOFF

__SOPT

__SRD

__SRW

__SSTR

__STDC_WANT_LIB_EXT1__

__SWR

__TVOS_9_0

__TVOS_9_1

__TVOS_9_2

__TVOS_10_0

__TVOS_10_0_1

__TVOS_10_1

__TVOS_10_2

__TVOS_11_0

__TVOS_11_1

__TVOS_11_2

__TVOS_11_3

__TVOS_11_4

__TVOS_12_0

__TVOS_12_1

__TVOS_12_2

__TVOS_12_3

__TVOS_12_4

__TVOS_13_0

__TVOS_13_2

__TVOS_13_3

__TVOS_13_4

__TVOS_14_0

__TVOS_14_1

__TVOS_14_2

__TVOS_14_3

__TVOS_14_5

__TVOS_14_6

__TVOS_14_7

__TVOS_15_0

__TVOS_15_1

__TVOS_15_2

__TVOS_15_3

__TVOS_15_4

__TVOS_15_5

__TVOS_15_6

__TVOS_16_0

__TVOS_16_1

__TVOS_16_2

__TVOS_16_3

__TVOS_16_4

__TVOS_16_5

__TVOS_16_6

__TVOS_17_0

__TVOS_17_1

__TVOS_17_2

__TVOS_17_3

__TVOS_17_4

__TVOS_17_5

__VISIONOS_1_0

__VISIONOS_1_1

__VISIONOS_1_2

__WATCHOS_1_0

__WATCHOS_2_0

__WATCHOS_2_1

__WATCHOS_2_2

__WATCHOS_3_0

__WATCHOS_3_1

__WATCHOS_3_1_1

__WATCHOS_3_2

__WATCHOS_4_0

__WATCHOS_4_1

__WATCHOS_4_2

__WATCHOS_4_3

__WATCHOS_5_0

__WATCHOS_5_1

__WATCHOS_5_2

__WATCHOS_5_3

__WATCHOS_6_0

__WATCHOS_6_1

__WATCHOS_6_2

__WATCHOS_7_0

__WATCHOS_7_1

__WATCHOS_7_2

__WATCHOS_7_3

__WATCHOS_7_4

__WATCHOS_7_5

__WATCHOS_7_6

__WATCHOS_8_0

__WATCHOS_8_1

__WATCHOS_8_3

__WATCHOS_8_4

__WATCHOS_8_5

__WATCHOS_8_6

__WATCHOS_8_7

__WATCHOS_8_8

__WATCHOS_9_0

__WATCHOS_9_1

__WATCHOS_9_2

__WATCHOS_9_3

__WATCHOS_9_4

__WATCHOS_9_5

__WATCHOS_9_6

__WATCHOS_10_0

__WATCHOS_10_1

__WATCHOS_10_2

__WATCHOS_10_3

__WATCHOS_10_4

__WATCHOS_10_5

__WORDSIZE

__bool_true_false_are_defined

__has_ptrcheck

__has_safe_buffers

false_

idtype_t_P_ALL

idtype_t_P_PGID

idtype_t_P_PID

mlx_array_dtype__MLX_BFLOAT16

mlx_array_dtype__MLX_BOOL

mlx_array_dtype__MLX_COMPLEX64

mlx_array_dtype__MLX_FLOAT16

mlx_array_dtype__MLX_FLOAT32

mlx_array_dtype__MLX_INT8

mlx_array_dtype__MLX_INT16

mlx_array_dtype__MLX_INT32

mlx_array_dtype__MLX_INT64

mlx_array_dtype__MLX_UINT8

mlx_array_dtype__MLX_UINT16

mlx_array_dtype__MLX_UINT32

mlx_array_dtype__MLX_UINT64

mlx_device_type__MLX_CPU

mlx_device_type__MLX_GPU

true_

Statics

__mb_cur_max

__stderrp

__stdinp

__stdoutp

suboptarg

sys_errlist

sys_nerr

Functions

_Exit⚠

__snprintf_chk⚠

__sprintf_chk⚠

__srget⚠

__svfscanf⚠

__swbuf⚠

__vsnprintf_chk⚠

__vsprintf_chk⚠

_mlx_error⚠

Throw an error.

a64l⚠

abort⚠

abs⚠

aligned_alloc⚠

alloca⚠

arc4random⚠

arc4random_addrandom⚠

arc4random_buf⚠

arc4random_stir⚠

arc4random_uniform⚠

asprintf⚠

atexit⚠

atexit_b⚠

atof⚠

atoi⚠

atol⚠

atoll⚠

bsearch⚠

bsearch_b⚠

calloc⚠

cgetcap⚠

cgetclose⚠

cgetent⚠

cgetfirst⚠

cgetmatch⚠

cgetnext⚠

cgetnum⚠

cgetset⚠

cgetstr⚠

cgetustr⚠

clearerr⚠

ctermid⚠

ctermid_r⚠

daemon⚠

devname⚠

devname_r⚠

div⚠

dprintf⚠

drand48⚠

ecvt⚠

erand48⚠

exit⚠

fclose⚠

fcvt⚠

fdopen⚠

feof⚠

ferror⚠

fflush⚠

fgetc⚠

fgetln⚠

fgetpos⚠

fgets⚠

fileno⚠

flockfile⚠

fmemopen⚠

fmtcheck⚠

fopen⚠

fprintf⚠

fpurge⚠

fputc⚠

fputs⚠

fread⚠

free⚠

freopen⚠

fscanf⚠

fseek⚠

fseeko⚠

fsetpos⚠

ftell⚠

ftello⚠

ftrylockfile⚠

funlockfile⚠

funopen⚠

fwrite⚠

gcvt⚠

getbsize⚠

getc⚠

getc_unlocked⚠

getchar⚠

getchar_unlocked⚠

getdelim⚠

getenv⚠

getiopolicy_np⚠

getline⚠

getloadavg⚠

getpriority⚠

getprogname⚠

getrlimit⚠

getrusage⚠

gets⚠

getsubopt⚠

getw⚠

grantpt⚠

heapsort⚠

heapsort_b⚠

initstate⚠

jrand48⚠

l64a⚠

labs⚠

lcong48⚠

ldiv⚠

llabs⚠

lldiv⚠

lrand48⚠

malloc⚠

malloc_type_aligned_alloc⚠

malloc_type_calloc⚠

malloc_type_free⚠

malloc_type_malloc⚠

malloc_type_posix_memalign⚠

malloc_type_realloc⚠

malloc_type_valloc⚠

malloc_type_zone_calloc⚠

malloc_type_zone_free⚠

malloc_type_zone_malloc⚠

malloc_type_zone_memalign⚠

malloc_type_zone_realloc⚠

malloc_type_zone_valloc⚠

mblen⚠

mbstowcs⚠

mbtowc⚠

mergesort⚠

mergesort_b⚠

mkstemp⚠

mktemp⚠

mlx_abs⚠

\defgroup ops Core array operations / /**@{

mlx_add⚠

mlx_addmm⚠

mlx_all_all⚠

mlx_all_axes⚠

mlx_all_axis⚠

mlx_allclose⚠

mlx_any⚠

mlx_any_all⚠

mlx_arange⚠

mlx_arccos⚠

mlx_arccosh⚠

mlx_arcsin⚠

mlx_arcsinh⚠

mlx_arctan⚠

mlx_arctan2⚠

mlx_arctanh⚠

mlx_argmax⚠

mlx_argmax_all⚠

mlx_argmin⚠

mlx_argmin_all⚠

mlx_argpartition⚠

mlx_argpartition_all⚠

mlx_argsort⚠

mlx_argsort_all⚠

mlx_array_data_bfloat16⚠

Returns a pointer to the array data, cast to bfloat16_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_bool⚠

Returns a pointer to the array data, cast to bool*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_complex64⚠

Returns a pointer to the array data, cast to _Complex*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_float16⚠

Returns a pointer to the array data, cast to float16_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_float32⚠

Returns a pointer to the array data, cast to float32*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_int8⚠

Returns a pointer to the array data, cast to int8_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_int16⚠

Returns a pointer to the array data, cast to int16_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_int32⚠

Returns a pointer to the array data, cast to int32_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_int64⚠

Returns a pointer to the array data, cast to int64_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_uint8⚠

Returns a pointer to the array data, cast to uint8_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_uint16⚠

Returns a pointer to the array data, cast to uint16_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_uint32⚠

Returns a pointer to the array data, cast to uint32_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_data_uint64⚠

Returns a pointer to the array data, cast to uint64_t*. Array must be evaluated, otherwise returns NULL.

mlx_array_dim⚠

The shape of the array in a particular dimension.

mlx_array_equal⚠

mlx_array_eval⚠

Evaluate the array.

mlx_array_from_bool⚠

New array from a bool scalar.

mlx_array_from_complex⚠

New array from a complex scalar.

mlx_array_from_data⚠

New array from existing buffer. @param data A buffer which will be copied. @param shape Shape of the array. @param dim Number of dimensions (size of shape). @param dtype Type of array elements.

mlx_array_from_float⚠

New array from a float scalar.

mlx_array_from_int⚠

New array from a int scalar.

mlx_array_get_dtype⚠

The array element type.

mlx_array_item_bfloat16⚠

Access the value of a scalar array.

mlx_array_item_bool⚠

Access the value of a scalar array.

mlx_array_item_complex64⚠

Access the value of a scalar array.

mlx_array_item_float16⚠

Access the value of a scalar array.

mlx_array_item_float32⚠

Access the value of a scalar array.

mlx_array_item_int8⚠

Access the value of a scalar array.

mlx_array_item_int16⚠

Access the value of a scalar array.

mlx_array_item_int32⚠

Access the value of a scalar array.

mlx_array_item_int64⚠

Access the value of a scalar array.

mlx_array_item_uint8⚠

Access the value of a scalar array.

mlx_array_item_uint16⚠

Access the value of a scalar array.

mlx_array_item_uint32⚠

Access the value of a scalar array.

mlx_array_item_uint64⚠

Access the value of a scalar array.

mlx_array_itemsize⚠

The size of the array’s datatype in bytes.

mlx_array_nbytes⚠

The number of bytes in the array.

mlx_array_ndim⚠

The array’s dimension.

mlx_array_shape⚠

The shape of the array. Returns: a pointer to the sizes of each dimension.

mlx_array_size⚠

Number of elements in the array.

mlx_array_strides⚠

The strides of the array.

mlx_as_strided⚠

mlx_astype⚠

mlx_async_eval⚠

\defgroup transforms Transform operations / /**@{

mlx_atleast_1d⚠

mlx_atleast_2d⚠

mlx_atleast_3d⚠

mlx_bitwise_and⚠

mlx_bitwise_or⚠

mlx_bitwise_xor⚠

mlx_block_masked_mm⚠

mlx_broadcast_arrays⚠

mlx_broadcast_to⚠

mlx_ceil⚠

mlx_checkpoint⚠

mlx_clip⚠

mlx_closure_apply⚠

Applies the closure over the given vector of arrays argument.

mlx_closure_new⚠

Returns a closure encapsulating the specified function.

mlx_closure_new_unary⚠

Returns a closure encapsulating the specified function. This is a convenience function, mapping mlx_array to mlx_vector_array under the hood.

mlx_closure_new_with_payload⚠

Returns a closure encapsulating the specified function, with the given payload. If dtor is not NULL, it will called when the closure is destroyed to free the payload.

mlx_closure_value_and_grad_apply⚠

Applies the closure over the given vector of arrays argument.

mlx_compile⚠

\defgroup compile Compilation operations / /**@{

mlx_concatenate⚠

mlx_concatenate_all⚠

mlx_conjugate⚠

mlx_conv1d⚠

mlx_conv2d⚠

mlx_conv3d⚠

mlx_conv_general⚠

mlx_copy⚠

mlx_cos⚠

mlx_cosh⚠

mlx_cpu_stream⚠

Returns the current default CPU stream.

mlx_cummax⚠

mlx_cummin⚠

mlx_cumprod⚠

mlx_cumsum⚠

mlx_default_device⚠

Returns the default MLX device.

mlx_default_stream⚠

Returns the default stream on the given device.

mlx_degrees⚠

mlx_depends⚠

mlx_dequantize⚠

mlx_detail_compile⚠

\defgroup transforms_impl Implementation detail operations / /**@{

mlx_detail_compile_clear_cache⚠

mlx_detail_compile_erase⚠

mlx_detail_vmap_replace⚠

mlx_detail_vmap_trace⚠

mlx_device_get_type⚠

Returns the type of the device.

mlx_device_new⚠

Returns a new device of specified type, with specified index.

mlx_diag⚠

mlx_diagonal⚠

mlx_disable_compile⚠

mlx_distributed_group_rank⚠

Get the rank.

mlx_distributed_group_size⚠

Get the group size.

mlx_distributed_group_split⚠

Split the group.

mlx_distributed_init⚠

Initialize distributed.

mlx_distributed_is_available⚠

Check if distributed is available.

mlx_divide⚠

mlx_divmod⚠

mlx_enable_compile⚠

mlx_equal⚠

mlx_erf⚠

mlx_erfinv⚠

mlx_eval⚠

mlx_exp⚠

mlx_expand_dims⚠

mlx_expm1⚠

mlx_eye⚠

mlx_fft_fft⚠

\defgroup fft FFT operations / /**@{

mlx_fft_fft2⚠

mlx_fft_fftn⚠

mlx_fft_ifft⚠

mlx_fft_ifft2⚠

mlx_fft_ifftn⚠

mlx_fft_irfft⚠

mlx_fft_irfft2⚠

mlx_fft_irfftn⚠

mlx_fft_rfft⚠

mlx_fft_rfft2⚠

mlx_fft_rfftn⚠

mlx_flatten⚠

mlx_floor⚠

mlx_floor_divide⚠

mlx_free⚠

Decreases the reference count of the object, and free if it gets to zero.

mlx_full⚠

mlx_future_wait⚠

mlx_gather⚠

mlx_gather_mm⚠

mlx_gather_qmm⚠

mlx_gpu_stream⚠

Returns the current default GPU stream.

mlx_greater⚠

mlx_greater_equal⚠

mlx_identity⚠

mlx_inner⚠

mlx_isclose⚠

mlx_isinf⚠

mlx_isnan⚠

mlx_isneginf⚠

mlx_isposinf⚠

mlx_jvp⚠

mlx_left_shift⚠

mlx_less⚠

mlx_less_equal⚠

mlx_linalg_cholesky⚠

\defgroup linalg Linear algebra operations / /**@{

mlx_linalg_inv⚠

mlx_linalg_norm⚠

mlx_linalg_norm_ord⚠

mlx_linalg_norm_p⚠

mlx_linalg_qr⚠

mlx_linalg_svd⚠

mlx_linspace⚠

mlx_load⚠

mlx_load_file⚠

\defgroup io IO operations / /**@{

mlx_load_safetensors⚠

mlx_load_safetensors_file⚠

mlx_log⚠

mlx_log1p⚠

mlx_log2⚠

mlx_log10⚠

mlx_logaddexp⚠

mlx_logical_and⚠

mlx_logical_not⚠

mlx_logical_or⚠

mlx_logsumexp⚠

mlx_logsumexp_all⚠

mlx_map_string_to_array_get⚠

Returns the value indexed at the specified key in the map. Returns NULL if no value was found for key.

mlx_map_string_to_array_insert⚠

Insert a new value at the specified key in the map. Returns true if the value was actually inserted.

mlx_map_string_to_array_iterate⚠

Returns a new iterator over the given map.

mlx_map_string_to_array_iterator_end⚠

Returns true iif iterator is at the end of the map.

mlx_map_string_to_array_iterator_key⚠

Returns the key associated to the current iterator position in the map.

mlx_map_string_to_array_iterator_next⚠

Increment iterator. Returns true if iterator could actually be incremented.

mlx_map_string_to_array_iterator_value⚠

Returns the value associated to the current iterator position in the map.

mlx_map_string_to_array_new⚠

Returns a new empty string-to-array map.

mlx_map_string_to_string_get⚠

Returns the value indexed at the specified key in the map. Returns NULL if no value was found for key.

mlx_map_string_to_string_insert⚠

Insert a new value at the specified key in the map. Returns true if the value was actually inserted.

mlx_map_string_to_string_iterate⚠

Returns a new iterator over the given map.

mlx_map_string_to_string_iterator_end⚠

Returns true iif iterator is at the end of the map.

mlx_map_string_to_string_iterator_key⚠

Returns the key associated to the current iterator position in the map.

mlx_map_string_to_string_iterator_next⚠

Increment iterator. Returns true if iterator could actually be incremented.

mlx_map_string_to_string_iterator_value⚠

Returns the value associated to the current iterator position in the map.

mlx_map_string_to_string_new⚠

Returns a new empty string-to-string map.

mlx_matmul⚠

mlx_max⚠

mlx_max_all⚠

mlx_maximum⚠

mlx_mean⚠

mlx_mean_all⚠

mlx_meshgrid⚠

mlx_metal_clear_cache⚠

\defgroup metal Metal specific operations / /**@{

mlx_metal_get_active_memory⚠

mlx_metal_get_cache_memory⚠

mlx_metal_get_peak_memory⚠

mlx_metal_is_available⚠

mlx_metal_reset_peak_memory⚠

mlx_metal_set_cache_limit⚠

mlx_metal_set_memory_limit⚠

mlx_metal_start_capture⚠

mlx_metal_stop_capture⚠

mlx_min⚠

mlx_min_all⚠

mlx_minimum⚠

mlx_moveaxis⚠

mlx_multiply⚠

mlx_negative⚠

mlx_not_equal⚠

mlx_number_of_elements⚠

mlx_ones⚠

mlx_ones_like⚠

mlx_outer⚠

mlx_pad⚠

mlx_partition⚠

mlx_partition_all⚠

mlx_power⚠

mlx_prod⚠

mlx_prod_all⚠

mlx_quantize⚠

mlx_quantized_matmul⚠

mlx_radians⚠

mlx_random_bernoulli⚠

\defgroup random Random number operations / /**@{

mlx_random_bits⚠

mlx_random_categorical⚠

mlx_random_categorical_num_samples⚠

mlx_random_categorical_shape⚠

mlx_random_gumbel⚠

mlx_random_key⚠

mlx_random_multivariate_normal⚠

mlx_random_normal⚠

mlx_random_randint⚠

mlx_random_seed⚠

mlx_random_split⚠

mlx_random_split_equal_parts⚠

mlx_random_truncated_normal⚠

mlx_random_uniform⚠

mlx_reciprocal⚠

mlx_remainder⚠

mlx_repeat⚠

mlx_repeat_all⚠

mlx_reshape⚠

mlx_retain⚠

Increases the reference count of the object.

mlx_right_shift⚠

mlx_round⚠

mlx_rsqrt⚠

mlx_safetensors_data⚠

Returns the data map.

mlx_safetensors_metadata⚠

Returns the metadata map.

mlx_safetensors_new⚠

Returns a empty safetensors storage.

mlx_save⚠

mlx_save_file⚠

mlx_save_safetensors⚠

mlx_save_safetensors_file⚠

mlx_scatter⚠

mlx_scatter_add⚠

mlx_scatter_max⚠

mlx_scatter_min⚠

mlx_scatter_prod⚠

mlx_set_default_device⚠

Set the default MLX device.

mlx_set_default_stream⚠

Set default stream.

mlx_set_error_handler⚠

Set the error handler.

mlx_sigmoid⚠

mlx_sign⚠

mlx_sin⚠

mlx_sinh⚠

mlx_slice⚠

mlx_slice_update⚠

mlx_softmax⚠

mlx_softmax_all⚠

mlx_sort⚠

mlx_sort_all⚠

mlx_split⚠

mlx_split_equal_parts⚠

mlx_sqrt⚠

mlx_square⚠

mlx_squeeze⚠

mlx_squeeze_all⚠

mlx_stack⚠

mlx_stack_all⚠

mlx_std⚠

mlx_std_all⚠

mlx_stop_gradient⚠

mlx_stream_equal⚠

Check if streams are the same.

mlx_stream_get_device⚠

Return the device of the stream.

mlx_stream_new⚠

Returns a new stream, with specified index, on a device.

mlx_stream_new_on_device⚠

Returns a new stream on a device.

mlx_string_data⚠

Returns a pointer to the string contents. The pointer is valid for the life duration of the string.

mlx_string_new⚠

Returns a new string, copying contents from str, which must end with \0.

mlx_subtract⚠

mlx_sum⚠

mlx_sum_all⚠

mlx_swapaxes⚠

mlx_synchronize⚠

Synchronize with the provided stream.

mlx_take⚠

mlx_take_all⚠

mlx_take_along_axis⚠

mlx_tan⚠

mlx_tanh⚠

mlx_tensordot⚠

mlx_tensordot_along_axis⚠

mlx_tile⚠

mlx_topk⚠

mlx_topk_all⚠

mlx_tostring⚠

Returns a new string representing the object.

mlx_trace⚠

mlx_transpose⚠

mlx_transpose_all⚠

mlx_tri⚠

mlx_tril⚠

mlx_triu⚠

mlx_value_and_grad⚠

mlx_var⚠

mlx_var_all⚠

mlx_vector_array_add⚠

Add an array to the vector of arrays. The reference count of the given array will be increased.

mlx_vector_array_add_arrays⚠

Add several arrays to the vector of arrays. The reference count of the given arrays will be increased.

mlx_vector_array_from_array⚠

Returns a new vector of array, containing a single specified array. The reference count of the given array will be increased.

mlx_vector_array_from_arrays⚠

Returns a new vector of arrays, from specified arrays. The reference count of the given arrays will be increased.

mlx_vector_array_get⚠

Get the array at index in the vector of arrays.

mlx_vector_array_new⚠

Returns a new empty vector of arrays

mlx_vector_array_size⚠

Size of the vector of arrays.

mlx_vector_vector_array_add⚠

Add an array vector to the vector of array vectors. Reference count of the array vector will be increased.

mlx_vector_vector_array_get⚠

Get the array vector at specified index in the vector of array vectors.

mlx_vector_vector_array_get2d⚠

Get the array at specified (index, arr_index) in the vector of array vectors.

mlx_vector_vector_array_new⚠

New empty vector of array vectors.

mlx_vector_vector_array_size⚠

mlx_view⚠

mlx_vjp⚠

mlx_where⚠

mlx_zeros⚠

mlx_zeros_like⚠

mrand48⚠

nrand48⚠

open_memstream⚠

pclose⚠

perror⚠

popen⚠

posix_memalign⚠

posix_openpt⚠

printf⚠

psort⚠

psort_b⚠

psort_r⚠

ptsname⚠

ptsname_r⚠

putc⚠

putc_unlocked⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

putw⚠

qsort⚠

qsort_b⚠

qsort_r⚠

radixsort⚠

rand⚠

rand_r⚠

random⚠

realloc⚠

reallocf⚠

realpath⚠

remove⚠

rename⚠

renameat⚠

renameatx_np⚠

renamex_np⚠

rewind⚠

rpmatch⚠

scanf⚠

seed48⚠

setbuf⚠

setbuffer⚠

setenv⚠

setiopolicy_np⚠

setkey⚠

setlinebuf⚠

setpriority⚠

setprogname⚠

setrlimit⚠

setstate⚠

setvbuf⚠

signal⚠

snprintf⚠

sprintf⚠

sradixsort⚠

srand⚠

srand48⚠

sranddev⚠

srandom⚠

srandomdev⚠

sscanf⚠

strtod⚠

strtof⚠

strtol⚠

strtold⚠

strtoll⚠

strtonum⚠

strtoq⚠

strtoul⚠

strtoull⚠

strtouq⚠

system⚠

tempnam⚠

tmpfile⚠

tmpnam⚠

ungetc⚠

unlockpt⚠

unsetenv⚠

valloc⚠

vasprintf⚠

vdprintf⚠

vfprintf⚠

vfscanf⚠

vprintf⚠

vscanf⚠

vsnprintf⚠

vsprintf⚠

vsscanf⚠

wait⚠

wait3⚠

wait4⚠

waitid⚠

waitpid⚠

wcstombs⚠

wctomb⚠

Type Aliases

FILE

__builtin_va_list

__darwin_blkcnt_t

__darwin_blksize_t

__darwin_clock_t

__darwin_ct_rune_t

__darwin_dev_t

__darwin_fsblkcnt_t

__darwin_fsfilcnt_t

__darwin_gid_t

__darwin_id_t

__darwin_ino64_t

__darwin_ino_t

__darwin_intptr_t

__darwin_mach_port_name_t

__darwin_mach_port_t

__darwin_mbstate_t

__darwin_mode_t

__darwin_natural_t

__darwin_nl_item

__darwin_off_t

__darwin_pid_t

__darwin_pthread_attr_t

__darwin_pthread_cond_t

__darwin_pthread_condattr_t

__darwin_pthread_key_t

__darwin_pthread_mutex_t

__darwin_pthread_mutexattr_t

__darwin_pthread_once_t

__darwin_pthread_rwlock_t

__darwin_pthread_rwlockattr_t

__darwin_pthread_t

__darwin_ptrdiff_t

__darwin_rune_t

__darwin_sigset_t

__darwin_size_t

__darwin_socklen_t

__darwin_ssize_t

__darwin_suseconds_t

__darwin_time_t

__darwin_uid_t

__darwin_useconds_t

__darwin_uuid_string_t

__darwin_uuid_t

__darwin_va_list

__darwin_wchar_t

__darwin_wctrans_t

__darwin_wctype_t

__darwin_wint_t

__int8_t

__int16_t

__int32_t

__int64_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint128_t

bfloat16_t

ct_rune_t

dev_t

float16_t

fpos_t

id_t

idtype_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

malloc_type_id_t

malloc_zone_t

mcontext_t

mlx_array

An N-dimensional array object.

mlx_array_dtype_

Array element type.

mlx_closure

A closure encapsulating a function taking a vector of arrays, and returning a vector of arrays

mlx_closure_value_and_grad

A closure which takes a vector of arrays, and returns a vector of vector of arrays of size 2.

mlx_device

A MLX device object.

mlx_device_type_

Device type.

mlx_distributed_group

A MLX distributed group object.

mlx_error_handler_func

\defgroup mlx_error Error management / /**@{

mlx_future

A MLX future object (pointing to a void).

mlx_map_string_to_array

A string-to-array map

mlx_map_string_to_array_iterator

An iterator over a string-to-array map.

mlx_map_string_to_string

A string-to-string map

mlx_map_string_to_string_iterator

An iterator over a string-to-string map.

mlx_safetensors

A MLX SafeTensors storage. Holds a string-to-array map (data) and string-to-string map (metadata).

mlx_stream

A MLX stream object.

mlx_string

A MLX string object.

mlx_vector_array

A vector of arrays.

mlx_vector_vector_array

A vector of array vectors.

mode_t

off_t

pid_t

pthread_attr_t

register_t

rlim_t

rune_t

rusage_info_current

rusage_info_t

sig_atomic_t

sig_t

siginfo_t

sigset_t

stack_t

syscall_arg_t

u_int8_t

u_int16_t

u_int32_t

u_int64_t

ucontext_t

uid_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

user_addr_t

user_long_t

user_off_t

user_size_t

user_ssize_t

user_time_t

user_ulong_t

va_list

wchar_t

Unions

__mbstate_t

__sigaction_u

sigval

wait