Module mlx_sys::ops::ffi

Functions

abs

add

addmm

Compute D = beta * C + alpha * (A @ B)

all

True if all elements in the std::unique_ptrmlx::core::array are true (or non-zero).

all_along_axes_keepdims

all_along_axis_keepdims

all_keepdims

True if all elements in the std::unique_ptrmlx::core::array are true (or non-zero).

allclose

True if the two arrays are equal within the specified tolerance.

any

any_along_axes_keepdims

any_along_axis_keepdims

any_keepdims

arange_f64

arange_i32

arange_start_stop_dtype

arange_start_stop_f64

arange_start_stop_i32

arange_stop_dtype

arange_stop_f64

arange_stop_i32

arccos

arccosh

arcsin

arcsinh

arctan

arctanh

argmax

argmax_along_axis_keepdims

argmax_keepdims

argmin

argmin_along_axis_keepdims

argmin_keepdims

argpartition

argpartition_along_axis

argsort

argsort_along_axis

array_equal

array_equal_equal_nan

True if two arrays have the same shape and elements.

as_strided

astype

broadcast_arrays

broadcast_to

ceil

clip

concatenate

concatenate_along_axis

conv1d

conv2d

copy

cos

cosh

cummax

cummin

cumprod

cumsum

depends

dequantize

diag

diagonal

divide

divmod

equal

erf

erfinv

exp

expand_dims_at_axes

expand_dims_at_axis

eye_n

eye_n_dtype

eye_n_m

eye_n_m_k

eye_n_m_k_dtype

flatten

flatten_start_axis_end_axis

floor

floor_divide

full_bfloat16_val

full_bfloat16_val_dtype

full_bool_val

full_bool_val_dtype

full_complex64_val

full_complex64_val_dtype

full_float16_val

full_float16_val_dtype

full_float32_val

full_float32_val_dtype

full_int8_val

full_int8_val_dtype

full_int16_val

full_int16_val_dtype

full_int32_val

full_int32_val_dtype

full_int64_val

full_int64_val_dtype

full_uint8_val

full_uint8_val_dtype

full_uint16_val

full_uint16_val_dtype

full_uint32_val

full_uint32_val_dtype

full_uint64_val

full_uint64_val_dtype

full_vals

full_vals_dtype

gather_along_axes

gather_along_axis

greater

greater_equal

identity

identity_dtype

inner

isclose

isinf

isnan

isneginf

isposinf

less

less_equal

linspace

log

log1p

log2

log10

logaddexp

logical_and

logical_not

logical_or

logsumexp

logsumexp_along_axes_keepdims

logsumexp_along_axis_keepdims

logsumexp_keepdims

matmul

max

max_along_axes_keepdims

max_along_axis_keepdims

max_keepdims

maximum

mean

mean_along_axes_keepdims

mean_along_axis_keepdims

mean_keepdims

min

min_along_axes_keepdims

min_along_axis_keepdims

min_keepdims

minimum

moveaxis

multiply

negative

not_equal

ones

ones_dtype

ones_like

outer

pad

pad_axes

pad_same_widths_for_each_axis

pad_unique_widths_for_each_axis

partition

partition_along_axis

power

Raise elements of a to the power of b element-wise

prod

prod_along_axes_keepdims

prod_along_axis_keepdims

prod_keepdims

quantize

quantized_matmul

reciprocal

remainder

repeat

repeat_along_axis

reshape

round

round_to_decimals

rsqrt

scatter_add_along_axes

scatter_add_along_axis

scatter_along_axes

scatter_along_axis

scatter_max_along_axes

scatter_max_along_axis

scatter_min_along_axes

scatter_min_along_axis

scatter_prod_along_axes

scatter_prod_along_axis

sigmoid

sign

sin

sinh

slice

slice_start_stop_strides

softmax

softmax_along_axes

softmax_along_axis

sort

sort_along_axis

split_at_indices

split_at_indices_along_axis

split_n

split_n_at_axis

sqrt

square

squeeze

squeeze_axes

squeeze_axis

stack

stack_along_axis

stop_gradient

subtract

sum

sum_along_axes_keepdims

sum_along_axis_keepdims

sum_keepdims

swapaxes

take

Take std::unique_ptrmlx::core::array slices at the given indices of the specified axis.

take_along_axis

take_flattened

Take std::unique_ptrmlx::core::array entries at the given indices treating the std::unique_ptrmlx::core::array as flattened.

tan

tanh

tensordot_list_dims

tensordot_ndims

tile

topk

topk_along_axis

transpose

transpose_axes

tri_n_dtype

tri_n_m_k_dtype

tril

triu

var

var_along_axes_keepdims

var_along_axis_keepdims

var_keepdims

where_condition

Renamed to where_condition because where is a reserved keyword in Rust.

zeros

zeros_dtype

zeros_like

Type Aliases

Dtype

OptionalArray

StreamOrDevice

array

bfloat16_t

complex64_t

float16_t